jest test config, module updates and test file changes
